+++ Changelog.md
@@ -1,3 +1,9 @@
+## 0.9.3
+Fix for attached token drag errors. #149.
+Fix display of SWADE cone template. #148.
+
+Update to libGeometry 0.4.6.
+
 ## 0.9.2
 Fix rotation setting being ignored. Issue #146.
 Add rotation for templates that are not centered on the attached token. Issue #145.